finalisation

+++ src/Controller/UserController.php
@@ -16,6 +16,8 @@ use App\Repository\UserRepository;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use JMS\Serializer\SerializerInterface;
+use App\EventSubscriber\ExceptionSubscriber;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class UserController extends AbstractController
 {
@@ -35,17 +37,21 @@ class UserController extends AbstractController
     }
 
     #[Route('/api/users/{userId}', name: 'api_get_user_details', methods: ['GET'])]
-    public function GetUsersDetails(CustomerRepository $customerRepository,UserRepository $userRepository,SerializerInterface $serializer,AuthorizationCheckerInterface $checkAuthorization, int $userId): JsonResponse
+    public function GetUsersDetails(CustomerRepository $customerRepository,UserRepository $userRepository,SerializerInterface $serializer,AuthorizationCheckerInterface $checkAuthorization, int $userId, ExceptionSubscriber $exception): JsonResponse
     {
         $customer = $this->getUser();
-
         $user = $userRepository->find($userId);
 
-        if ($customer == $user->getCustomer()){
+        if ($user === null){
+            $data = [
+                'status' => "404",
+                'message' => "L'utilisateur n'existe pas en base de donnée"
+            ];
+            return new JsonResponse($data);
+        }
+        if ($customer === $user->getCustomer()){
             $context  = SerializationContext::create()->setGroups(['ShowUsers','ShowUsersDetails']);
-            $context1 = SerializationContext::create()->setGroups(['ShowUsersDetails']);
             $jsonUserDetails = $serializer->serialize($user, 'json', $context);
-
 
             return new JsonResponse(
                 $jsonUserDetails, Response::HTTP_OK, [], true,
@@ -63,31 +69,45 @@ class UserController extends AbstractController
         $customer = $this->getUser();
         $user = $userRepository->find($userId);
 
+        if ($user === null){
+            $data = [
+                'status' => "404",
+                'message' => "L'utilisateur n'existe pas en base de donnée"
+            ];
+            return new JsonResponse($data);
+        }
+
             if ($customer === $user->getCustomer()) {
 
                 $em->remove($user);
                 $em->flush();
 
                 return new JsonResponse(null, Response::HTTP_NO_CONTENT);
-
             } else {
-
                 return new JsonResponse(Response::HTTP_BAD_REQUEST);
             }
     }
 
     #[Route('/api/user', name: 'api_post_user', methods: ['POST'])]
-    public function AddUser(Request $request, SerializerInterface $serializer, EntityManagerInterface $entityManager, UrlGeneratorInterface $urlGenerator,CustomerRepository $customerRepository): JsonResponse
+    public function AddUser(Request $request, SerializerInterface $serializer, EntityManagerInterface $entityManager, UrlGeneratorInterface $urlGenerator,CustomerRepository $customerRepository, ValidatorInterface $validator): JsonResponse
     {
         $customer = $this->getUser();
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
         $user->setCreatedAt(new \DateTimeImmutable());
         $user->setCustomer($customer);
 
+        // On vérifie les erreurs
+        $errors = $validator->validate($user);
+
+        if ($errors->count() > 0) {
+            return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
+        }
+
         $entityManager->persist($user);
         $entityManager->flush();
+        $context  = SerializationContext::create()->setGroups(['ShowUsers','ShowUsersDetails']);
 
-        $jsonUserDetails = $serializer->serialize($user, 'json', ['groups' => ['ShowUsers','ShowUserDetails']]);
+        $jsonUserDetails = $serializer->serialize($user, 'json', $context);
 
         return new JsonResponse($jsonUserDetails, Response::HTTP_OK, [], true);
